pyproject.toml
[tool.poetry]
name = "seo-title-generator-st"
version = "0.0.1"
description = ""
authors = ["sneakyPad"]
readme = "README.md"
packages = [{include = "seo_title_generator_st"}]

[tool.poetry.dependencies]
python = "3.9.17"
streamlit = "^1.23.1"
requests = "^2.31.0"
streamlit-analytics = "^0.4.1"
st-annotated-text = "^4.0.0"
hugchat = "^0.0.8"
retry = "^0.9.2"
streamlit-extras = "^0.2.7"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
